Fleshed out the race list
Decided on a final look and made empty pages for all the rest of the races. I've started bringing the existing races up to the current visual look. Also settled on how the legacy and current base race will be visualized Firblog is the only one that has it represented so far.